impl_for_type -> PrimitiveType::impls
